requirements.txt
Kivy==2.2.1
python-dotenv==1.0.0
kivymd==1.1.1
